refactor: utils-sign.js 내용을 app.js로 이동 	new file:   scripts/userdb.js 	deleted:    scripts/utils-sign.js 	modified:   signin/app.js 	modified:   signup/app.js
