+++ scripts/userdb.js
@@ -1,0 +1,9 @@
+/* USER DATABASE */
+
+//유효한 로그인 정보
+const VALID_USER = {
+  email: "test@codeit.com",
+  password: "codeit101",
+};
+
+export { VALID_USER };
