+++ signin/app.js
@@ -15,7 +15,26 @@ import {
   hidePassword,
 } from "../scripts/ui-sign.js";
 
-import { isValidEmailFormat, getUserByEmail } from "../scripts/utils-sign.js";
+import { isValidEmailFormat, getUserByEmail } from "../scripts/userdb.js";
+
+/********************
+ * UTILITY FUNCTION
+ ********************/
+
+//이메일 양식 유효성 검사
+function isValidEmailFormat(emailInputValue) {
+  const regex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
+  return regex.test(emailInputValue);
+}
+
+//이메일이 일치하는 유저 데이터 반환
+function getUserByEmail(emailInputValue) {
+  if (VALID_USER.email !== emailInputValue) {
+    return null;
+  }
+
+  return VALID_USER;
+}
 
 /********************
  * ACTIVE FUNCTION
